# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Box API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authenticated account
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUser'}
...
  /auth/change-password:
    post:
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangePasswordRequest'}
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUser'}
  /members:
    get:
      operationId: listMembers
      responses:
        '200':
          description: Organization accounts and external identities
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Member'}
    post:
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAccountRequest'}
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
  /members/{memberId}:
    parameters:
      - name: memberId
        in: path
        required: true
        schema: {type: string, format: uuid}
    patch:
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateAccountRequest'}
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
  /members/{memberId}/password:
    parameters:
      - name: memberId
        in: path
        required: true
        schema: {type: string, format: uuid}
    put:
      operationId: resetAccountPassword
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResetPasswordRequest'}
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
...
  /agents:
    get:
      operationId: listAgents
      responses:
        '200':
          description: Agent definitions
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Agent'}
    post:
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAgentRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Agent'}
...
  /workspaces:
    get:
      operationId: listWorkspaces
      responses:
        '200':
          description: Workspaces
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Workspace'}
    post:
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateWorkspaceRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Workspace'}
  /boxes:
    get:
      operationId: listBoxes
      responses:
        '200':
          description: Boxes
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Box'}
    post:
      operationId: createBox
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateBoxRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Box'}
  /boxes/{boxId}:
    parameters:
      - {$ref: '#/components/parameters/BoxId'}
    get:
      operationId: getBox
      responses:
        '200':
          description: Box snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoxSnapshot'}
  /boxes/{boxId}/messages:
    parameters:
      - {$ref: '#/components/parameters/BoxId'}
    get:
      operationId: listMessages
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Message'}
    post:
      operationId: sendMessage
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendMessageRequest'}
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Message'}
  /boxes/{boxId}/events:
    parameters:
      - {$ref: '#/components/parameters/BoxId'}
    get:
      operationId: streamBoxEvents
      parameters:
        - name: Last-Event-ID
          in: header
          schema: {type: integer, format: int64}
      responses:
        '200':
          description: Server-sent event stream
          content:
            text/event-stream:
              schema: {type: string}
  /boxes/{boxId}/interrupt:
    parameters:
      - {$ref: '#/components/parameters/BoxId'}
    post:
      operationId: interruptBox
      responses:
        '202': {description: Accepted}
  /boxes/{boxId}/stop:
    parameters:
      - {$ref: '#/components/parameters/BoxId'}
    post:
      operationId: stopBox
      responses:
        '202': {description: Accepted}
  /boxes/{boxId}/resume:
    parameters:
      - {$ref: '#/components/parameters/BoxId'}
    post:
      operationId: resumeBox
      responses:
        '202': {description: Accepted}
...
  /approvals/{approvalId}/decision:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema: {type: string, format: uuid}
    post:
      operationId: decideApproval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: {enum: [approved, denied]}
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: agentbox_session
  parameters:
    BoxId:
      name: boxId
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    CurrentUser:
      type: object
      required: [id, login, displayName, role, status, mustChangePassword]
      properties:
        id: {type: string, format: uuid}
        login: {type: string}
        displayName: {type: string}
        role: {enum: [admin, user]}
        status: {enum: [active, disabled]}
        mustChangePassword: {type: boolean}
    Member:
      allOf:
        - {$ref: '#/components/schemas/CurrentUser'}
        - type: object
          required: [organizationId, hasPassword, createdAt]
          properties:
            organizationId: {type: string, format: uuid}
            hasPassword: {type: boolean}
            createdAt: {type: string, format: date-time}
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 64}
        password: {type: string, minLength: 8, maxLength: 72}
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string}
        newPassword: {type: string, minLength: 8, maxLength: 72}
    CreateAccountRequest:
      type: object
      required: [username, displayName, role, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 64}
        displayName: {type: string, minLength: 1, maxLength: 128}
        role: {enum: [admin, user]}
        password: {type: string, minLength: 8, maxLength: 72}
    UpdateAccountRequest:
      type: object
      properties:
        displayName: {type: string, minLength: 1, maxLength: 128}
        role: {enum: [admin, user]}
        status: {enum: [active, disabled]}
    ResetPasswordRequest:
      type: object
      required: [password]
      properties:
        password: {type: string, minLength: 8, maxLength: 72}
    Agent:
      type: object
      required: [id, name, runtimeType, version]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        runtimeType: {enum: [omp, claude, acp]}
        version: {type: integer}
        model: {type: [string, 'null']}
        systemPrompt: {type: string}
    CreateAgentRequest:
      type: object
      required: [name, runtimeType]
      properties:
        name: {type: string}
        runtimeType: {enum: [omp, claude, acp]}
        model: {type: [string, 'null']}
        systemPrompt: {type: string}
...
    Workspace:
      type: object
      required: [id, hostId, name, path, status]
      properties:
        id: {type: string, format: uuid}
        hostId: {type: string, format: uuid}
        name: {type: string}
        path: {type: string}
        status: {enum: [provisioning, ready, error, archived]}
    CreateWorkspaceRequest:
      type: object
      required: [hostId, name, path]
      properties:
        hostId: {type: string, format: uuid}
        name: {type: string}
        path: {type: string}
    Box:
      type: object
      required: [id, name, agentId, hostId, workspaceId, status]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        agentId: {type: string, format: uuid}
        hostId: {type: string, format: uuid}
        workspaceId: {type: string, format: uuid}
        status: {type: string}
        runtimeType: {type: string}
        updatedAt: {type: string, format: date-time}
    CreateBoxRequest:
      type: object
      required: [name, agentId, hostId, workspaceId]
      properties:
        name: {type: string}
        agentId: {type: string, format: uuid}
        hostId: {type: string, format: uuid}
        workspaceId: {type: string, format: uuid}
    BoxSnapshot:
      allOf:
        - {$ref: '#/components/schemas/Box'}
        - type: object
          properties:
            activeRunId: {type: [string, 'null'], format: uuid}
            lastEventSeq: {type: integer, format: int64}
    Message:
      type: object
      required: [id, boxId, role, status, content, createdAt]
      properties:
        id: {type: string, format: uuid}
        boxId: {type: string, format: uuid}
        runId: {type: [string, 'null'], format: uuid}
        authorName: {type: [string, 'null']}
        role: {enum: [user, assistant, system]}
        delivery: {type: [string, 'null'], enum: [prompt, steer, follow_up, null]}
        status: {type: string}
        content: {type: string}
        createdAt: {type: string, format: date-time}
    SendMessageRequest:
      type: object
      required: [content, delivery]
      properties:
        content: {type: string, minLength: 1}
        delivery: {enum: [prompt, steer, follow_up]}
    Approval:
      type: object
      required: [id, boxId, runId, toolName, riskLevel, status, expiresAt]
      properties:
        id: {type: string, format: uuid}
        boxId: {type: string, format: uuid}
        runId: {type: string, format: uuid}
        toolName: {type: string}
        riskLevel: {enum: [low, medium, high, critical]}
        status: {enum: [pending, approved, denied, expired, cancelled]}
        payload: {type: object}
        expiresAt: {type: string, format: date-time}